code/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
python-docx>=1.1.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0

# Data processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
recordlinkage>=0.16.0

# PII/PHI detection and redaction
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.5.0

# Database connectors
psycopg2-binary>=2.9.0
pymysql>=1.1.0
pymongo>=4.6.0
cx_Oracle>=8.3.0

# File format handlers
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10
tomli>=2.0.0

# Enterprise system connectors (optional)
simple-salesforce>=1.12.0
hubspot-api-client>=7.0.0
msal>=1.24.0
google-api-python-client>=2.100.0

# Cloud and scalability
boto3>=1.28.0
dask>=2023.8.0
ray>=2.6.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

